+++ projects/ez-vid-ang/src/lib/core/directives/tooltip.spec.ts
@@ -1,0 +1,7 @@
+import { EvaTooltip } from './tooltip';
+
+describe('EvaTooltip', () => {
+  it('should be defined', () => {
+    expect(EvaTooltip).toBeDefined();
+  });
+});